Moved some tests from test_onnx.rs
