+++ crates/burn-import/onnx-tests/tests/test_mod.rs
@@ -7,5 +7,11 @@ mod node_tests;
 
 // Import individual node modules
 pub mod add;
+pub mod constant;
+pub mod constant_of_shape;
+pub mod div;
+pub mod matmul;
+pub mod mean;
+pub mod mul;
 pub mod sub;
 pub mod sum;
